mmaprototype: move MMARebalanceAdvisor to rebalance_advisor.go
This commit moves `MMARebalanceAdvisor` related code from `allocator_state.go` to
`rebalance_advisor.go`.
